+++ src/main/java/com/rkmd/toki_no_nagare/controller/payment/PaymentController.java
@@ -1,6 +1,8 @@
 package com.rkmd.toki_no_nagare.controller.payment;
 
 import com.rkmd.toki_no_nagare.dto.payment.*;
+import com.rkmd.toki_no_nagare.exception.UnAuthorizedException;
+import com.rkmd.toki_no_nagare.service.AuthorizationService;
 import com.rkmd.toki_no_nagare.service.PaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,10 +15,18 @@ public class PaymentController implements PaymentControllerResources{
   @Autowired
   private PaymentService paymentService;
 
+  @Autowired
+  private AuthorizationService authorizationService;
+
   @PutMapping(value = "/v1/payments", produces = "application/json")
   @ResponseStatus(value = HttpStatus.OK)
+  //public ChangePaymentResponseDto changePaymentStatus(ChangePaymentRequestDto request, @RequestHeader("x-auth-username") String userName, @RequestHeader("x-auth-password") String password) {
   public ChangePaymentResponseDto changePaymentStatus(ChangePaymentRequestDto request) {
-    return paymentService.changePaymentStatus(request.getBookingCode(), request.getPaymentStatus());
+    //if (!authorizationService.validatePassword(userName, password))
+      //throw new UnAuthorizedException("invalid_password", "The password is invalid");
+
+    String userName = "SOME_USER";
+    return paymentService.changePaymentStatus(request.getBookingCode(), request.getPaymentStatus(), userName);
   }
 
   @GetMapping(value = "/v1/payments", produces = "application/json")