+++ src/main/java/com/rkmd/toki_no_nagare/dto/user/CreateUserResponseDto.java
@@ -1,6 +1,5 @@
 package com.rkmd.toki_no_nagare.dto.user;
 
-import com.rkmd.toki_no_nagare.entities.user.RoleType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 
@@ -13,11 +12,5 @@ public class CreateUserResponseDto {
       description = "detail: Nombre de usuario",
       example = "rodri_bangu")
   public String userName;
-
-  @Schema(
-      name = "role",
-      description = "detail: Rol del usuario",
-      example = "admin")
-  public RoleType role;
 
 }