+++ src/main/java/com/rkmd/toki_no_nagare/dto/user/GetUserResponseDto.java
@@ -1,6 +1,5 @@
 package com.rkmd.toki_no_nagare.dto.user;
 
-import com.rkmd.toki_no_nagare.entities.user.RoleType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
@@ -16,11 +15,4 @@ public class GetUserResponseDto {
       description = "detail: Nombre de usuario",
       example = "tiki_taka")
   private String userName;
-
-  @Schema(
-      name = "role",
-      description = "detail: Role del usuario",
-      example = "viewer")
-  private RoleType role;
-
 }