+++ src/main/java/com/rkmd/toki_no_nagare/exception/UnAuthorizedException.java
@@ -1,0 +1,9 @@
+package com.rkmd.toki_no_nagare.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class UnAuthorizedException extends ApiException {
+    public UnAuthorizedException(String code, String description) {
+        super(code, description, HttpStatus.UNAUTHORIZED);
+    }
+}