+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
+### 2023.09.08
+- add: prereserve validation before prereserving
+
 ### 2023.09.07
+- delete: RoleType 
+- delete: relation between User and Contact
+- add: /v1/login 
+- refactor: clean unused endpoints
+- add: seller to ChangePaymentResponseDto 
+- add: seller AUTOMATIC_EXPIRATION_JOB 
+- add: authorization for changePaymentStatus()
+- add: UnAuthorizedException
+- add: validStatusTransition for booking, payment and seat
 - refactor: Se modifica la implementación de los reportes estadisticos (0edd119)
 - refactor: Se modifica tipo de respuesta en el endpoint de reportes estadisticos (d8a691c)
 - refactor: Se agregan clases DTO para las respuestas de los endpoints de reporte (cae5aa5)